+++ src/main/java/de/terrarier/netlistening/api/type/DataTypeByteArray.java
@@ -14,7 +14,7 @@ import org.jetbrains.annotations.NotNull;
 public final class DataTypeByteArray extends DataType<byte[]> {
 	
 	protected DataTypeByteArray() {
-		super((byte) 0x3, 4, true);
+		super((byte) 0x3, (byte) 4, true);
 	}
 
 	@Override
