+++ src/main/java/de/terrarier/netlistening/api/type/DataTypeLong.java
@@ -12,7 +12,7 @@ import org.jetbrains.annotations.NotNull;
 public final class DataTypeLong extends DataType<Long> {
 	
 	protected DataTypeLong() {
-		super((byte) 0x9, 8, true);
+		super((byte) 0x9, (byte) 8, true);
 	}
 
 	@Override
