+++ src/main/java/de/terrarier/netlistening/api/type/DataTypeObject.java
@@ -16,7 +16,7 @@ import org.jetbrains.annotations.NotNull;
 public final class DataTypeObject extends DataType<Object> {
 	
 	protected DataTypeObject() {
-		super((byte) 0x8, 4, true);
+		super((byte) 0x8, (byte) 4, true);
 	}
 
 	@Override
